fix: class services need to be in sibling files for server client separation
