+++ apps/webapp/app/routes/api.v1.runs.$runId.tasks/route.ts
@@ -1,17 +1,11 @@
 import type { ActionFunctionArgs } from "@remix-run/server-runtime";
 import { json } from "@remix-run/server-runtime";
-import {
-  API_VERSIONS,
-  RunTaskBodyOutputSchema,
-  RunTaskResponseWithCachedTasksBody,
-  ServerTask,
-} from "@trigger.dev/core";
+import { API_VERSIONS, RunTaskBodyOutputSchema } from "@trigger.dev/core";
 import { z } from "zod";
-import { PrismaClient, prisma } from "~/db.server";
-import { prepareTasksForCaching } from "~/models/task.server";
 import { authenticateApiRequest } from "~/services/apiAuth.server";
 import { logger } from "~/services/logger.server";
 import { RunTaskService } from "~/services/tasks/runTask.server";
+import { ChangeRequestLazyLoadedCachedTasks } from "./ChangeRequestLazyLoadedCachedTasks";
 
 const ParamsSchema = z.object({
   runId: z.string(),
@@ -109,50 +103,5 @@ export async function action({ request, params }: ActionFunctionArgs) {
     }
 
     return json({ error: "Something went wrong" }, { status: 500 });
-  }
-}
-
-class ChangeRequestLazyLoadedCachedTasks {
-  #prismaClient: PrismaClient;
-
-  constructor(prismaClient: PrismaClient = prisma) {
-    this.#prismaClient = prismaClient;
-  }
-
-  public async call(
-    runId: string,
-    task: ServerTask,
-    cursor?: string | null
-  ): Promise<RunTaskResponseWithCachedTasksBody> {
-    if (!cursor) {
-      return {
-        task,
-      };
-    }
-
-    // We need to limit the cached tasks to not be too large >2MB when serialized
-    const TOTAL_CACHED_TASK_BYTE_LIMIT = 2000000;
-
-    const nextTasks = await this.#prismaClient.task.findMany({
-      where: {
-        runId,
-        status: "COMPLETED",
-        noop: false,
-      },
-      take: 250,
-      cursor: {
-        id: cursor,
-      },
-      orderBy: {
-        id: "asc",
-      },
-    });
-
-    const preparedTasks = prepareTasksForCaching(nextTasks, TOTAL_CACHED_TASK_BYTE_LIMIT);
-
-    return {
-      task,
-      cachedTasks: preparedTasks,
-    };
   }
 }
